Split sample loading helpers and clear clippy debt
