+++ src/gui_app/waveform/interaction/edit_fade_drag.rs
@@ -20,30 +20,30 @@ impl WaveformEditFadeDrag {
         selection: wavecrate::selection::SelectionRange,
     ) -> Self {
         let curve = match handle {
-            WaveformEditFadeHandle::FadeInEnd
-            | WaveformEditFadeHandle::FadeInStart
-            | WaveformEditFadeHandle::FadeInOuterStart => {
+            WaveformEditFadeHandle::InEnd
+            | WaveformEditFadeHandle::InStart
+            | WaveformEditFadeHandle::InOuterStart => {
                 selection.fade_in().map(|fade| fade.curve).unwrap_or(0.5)
             }
-            WaveformEditFadeHandle::FadeOutStart
-            | WaveformEditFadeHandle::FadeOutEnd
-            | WaveformEditFadeHandle::FadeOutOuterEnd => {
+            WaveformEditFadeHandle::OutStart
+            | WaveformEditFadeHandle::OutEnd
+            | WaveformEditFadeHandle::OutOuterEnd => {
                 selection.fade_out().map(|fade| fade.curve).unwrap_or(0.5)
             }
         };
         let fixed_ratio = match handle {
-            WaveformEditFadeHandle::FadeInStart => selection
+            WaveformEditFadeHandle::InStart => selection
                 .fade_in()
                 .map(|fade| selection.start() + selection.width() * fade.length)
                 .unwrap_or(selection.start()),
-            WaveformEditFadeHandle::FadeOutEnd => selection
+            WaveformEditFadeHandle::OutEnd => selection
                 .fade_out()
                 .map(|fade| selection.end() - selection.width() * fade.length)
                 .unwrap_or(selection.end()),
-            WaveformEditFadeHandle::FadeInEnd
-            | WaveformEditFadeHandle::FadeOutStart
-            | WaveformEditFadeHandle::FadeInOuterStart
-            | WaveformEditFadeHandle::FadeOutOuterEnd => 0.0,
+            WaveformEditFadeHandle::InEnd
+            | WaveformEditFadeHandle::OutStart
+            | WaveformEditFadeHandle::InOuterStart
+            | WaveformEditFadeHandle::OutOuterEnd => 0.0,
         };
         Self {
             handle,
@@ -60,22 +60,20 @@ impl WaveformEditFadeDrag {
     ) -> wavecrate::selection::SelectionRange {
         let ratio = ratio.clamp(0.0, 1.0);
         match self.handle {
-            WaveformEditFadeHandle::FadeInEnd => {
+            WaveformEditFadeHandle::InEnd => {
                 resize_fade_in_end_with_collision(selection, self.baseline, ratio, self.curve)
             }
-            WaveformEditFadeHandle::FadeOutStart => {
+            WaveformEditFadeHandle::OutStart => {
                 resize_fade_out_start_with_collision(selection, self.baseline, ratio, self.curve)
             }
-            WaveformEditFadeHandle::FadeInStart => {
+            WaveformEditFadeHandle::InStart => {
                 resize_fade_in_start(self.baseline, self.fixed_ratio, ratio, self.curve)
             }
-            WaveformEditFadeHandle::FadeOutEnd => {
+            WaveformEditFadeHandle::OutEnd => {
                 resize_fade_out_end(self.baseline, self.fixed_ratio, ratio, self.curve)
             }
-            WaveformEditFadeHandle::FadeInOuterStart => {
-                resize_fade_in_outer_start(selection, ratio)
-            }
-            WaveformEditFadeHandle::FadeOutOuterEnd => resize_fade_out_outer_end(selection, ratio),
+            WaveformEditFadeHandle::InOuterStart => resize_fade_in_outer_start(selection, ratio),
+            WaveformEditFadeHandle::OutOuterEnd => resize_fade_out_outer_end(selection, ratio),
         }
     }
 }
